Fix incorrect result for consecutive operation
When performing consecutive operation without clicking equal sign, the
calculation result is based on the last operation only instead of the
whole operation.

Cause of the issue is the operation is only performed when equal sign is
clicked and the num1 is renew as the number before the operator every
time when operator is clicked. Thus, the calculation is based on the
last operation.

Therefore, applied fix to calculate each intermediate result for
consecutive operation.